// FashionShop.Web/src/features/shop/coupons/types/coupon.ts
export interface Coupon {
    id: string;
    name: string;
    code: string;
    description: string;
    discountType: DiscountType;
    discountAmount: number;
    maxDiscountAmount?: number | null;
    minOrderValue: number;
    quantity: number;
    usedCount: number;
    remainingUsagePerUser: number;
    startDate: string;
    endDate: string;
}

export const DiscountType = {
    FixedAmount: "FixedAmount",
    Percentage: "Percentage",
} as const;

export type DiscountType = typeof DiscountType[keyof typeof DiscountType];